feat(activities): reuse activity cards across overview and planning
Extract reusable activity card, searchable list helpers, and a category-filtered suggestions panel. Add an accommodation suggestions tab in trip overview while keeping existing activity voting and detail editing flows.
