+++ lib/l10n/app_en.arb
@@ -855,6 +855,10 @@
   "tripOverviewTileComingSoon": "[Coming soon]",
   "tripOverviewMyRoom": "My room",
   "tripOverviewMyRooms": "My rooms",
+  "tripOverviewTabSummary": "Overview",
+  "tripOverviewTabAccommodationSuggestions": "Accommodation",
+  "tripOverviewNoAccommodationSuggestions": "No accommodation suggestions yet. Add an \"Accommodation\" activity here or from Planning to make it appear.",
+  "tripOverviewNoRestaurantSuggestions": "No restaurant suggestions yet. Add a \"Restaurant\" activity here or from Planning to make it appear.",
   "cupidonPopupTitle": "You have a match",
   "cupidonPopupViewMatchesAction": "See my matches",
   "cupidonPopupUnknownMember": "Someone",
