+++ lib/l10n/app_fr.arb
@@ -855,6 +855,10 @@
   "tripOverviewTileComingSoon": "[À venir]",
   "tripOverviewMyRoom": "Ma chambre",
   "tripOverviewMyRooms": "Mes chambres",
+  "tripOverviewTabSummary": "Aperçu",
+  "tripOverviewTabAccommodationSuggestions": "Hébergements",
+  "tripOverviewNoAccommodationSuggestions": "Aucune suggestion d'hébergement. Ajoute une activité \"Hébergement\" ici ou depuis le planning pour la voir apparaître.",
+  "tripOverviewNoRestaurantSuggestions": "Aucune suggestion de restaurant. Ajoute une activité \"Restaurant\" ici ou depuis le planning pour la voir apparaître.",
   "cupidonPopupTitle": "Tu as un match",
   "cupidonPopupViewMatchesAction": "Voir mes matchs",
   "cupidonPopupUnknownMember": "Quelqu'un",
